// src/types.ts
export type Car = {
  id: number;
  brand: string;
  model: string;
  year: number;
  price: number;
  mileage: number;
  fuelType: string;
  transmission: string;
  horsepower: number;
  torque: number;
  acceleration: number;
  topSpeed: number;
  fuelConsumption: number;
  co2: number;
  rating: number;
  image: string;
  features: string[];
};